change structure to fit general rule
